Keep: [U済] IsFilePath()整理
[patches:#873]


git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk@4209 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/file.h
@@ -30,7 +30,7 @@
 
 bool fexist(LPCTSTR pszPath); //!< ファイルまたはディレクトリが存在すればtrue
 
-bool IsFilePath( const char*, int*, int*, bool = true );
+bool IsFilePath( const char*, size_t*, size_t*, bool = true );
 bool IsFileExists(const TCHAR* path, bool bFileOnly = false);
 
 //	Apr. 30, 2003 genta
